fix: address browser screenshot review feedback
